Update EditCustomerForm.cs (#1)
* Update EditCustomerForm.cs

The current version of the form calls the NextCustomerId method inside CreateCustomerLoad().
This causes that every time you click the "Create User" button, even if you don’t actually create a new user and just close the EditCustomerForm, a new customer ID is generated and never used.

To avoid this and prevent "lost" IDs, it's better to generate the new ID from the last item in the list using .Last(). This way, the new ID is always based on the latest existing one.

Changes made:

- Updated the CreateCustomerLoad() method in EditCustomerForm.cs.

P.S.: The RentalHistoryForm.cs had a "Close" button with no functionality. It's necessary to implement Dispose().

* Update RentalHistoryForm.cs

Add Dispose () to "Close" button.

* Update2 EditCustomerForm.cs

I set the label and the textBox with the property visible = false.
I set the textBoxId to 0.

I add the NextCustomerId method into CreateCustomer() to only apply when we're sure we're creating a new customer.